% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{add.messages}
\alias{add.messages}
\title{Add chat messages for chatGPT}
\usage{
\method{add}{messages}(messages, assistant = NULL, user = NULL, ...)
}
\arguments{
\item{messages}{messages. chatGPT와 chat completion을 수행하기 위한 메시지 객체.}

\item{assistant}{character. assistant role을 갖는 메시지.}

\item{user}{character. user role을 갖는 메시지.}
}
\value{
messages object.
}
\description{
chatGPT의 chat completion messages에 메시지를 추가
}
\examples{
\dontrun{
# 메시지 생성
msg <- create_messages("Who won the world series in 2020?", "You are a helpful assistant.")
msg

# 메시지 추가
library(dplyr)

msg <- msg \%>\%
  add(assistant = "The Los Angeles Dodgers won the World Series in 2020.",
      user = "Where was it played?")
msg

}
}
