% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitGPT.R
\docType{package}
\name{bitGPT-package}
\alias{bitGPT}
\alias{bitGPT-package}
\title{Tools to use chatGPT for Koreans}
\description{
A collection of tools that support using chatGPT for Korean.
}
\details{
\tabular{ll}{
Package: \tab bitGPT\cr
Type: \tab Package\cr
Version: \tab 0.3.6.\cr
Date: \tab 2023-04-02\cr
License: \tab GPL (>= 3)\cr
}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/bit2r/bitGPT/issues}
}

}
\author{
유충현
Maintainer: 유충현 \href{mailto:choonghyun.ryu@gmail.com}{choonghyun.ryu@gmail.com}
}
