% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chat_completion}
\alias{chat_completion}
\title{Chat completion with chatGPT}
\usage{
chat_completion(
  messages = NULL,
  model = c("gpt-3.5-turbo", "gpt-3.5-turbo-0301"),
  temperature = 1,
  top_p = 1,
  type = c("text", "console", "viewer"),
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{messages}{character. 채팅을 위한 메시지. 메시지는 영문과 국문 모두 가능함.}

\item{model}{character. Chat completion에 사용할 OpenAI의 모델로,
"gpt-3.5-turbo", "gpt-3.5-turbo-0301"에서 선택함. 기본값은 "gpt-3.5-turbo".}

\item{temperature}{numeric. 0에서 2 사이에서 사용할 샘플링 온도.
0.8과 같이 값이 높으면 출력이 더 무작위적이고,
0.2와 같이 값이 낮으면 더 집중적이고 결정론적인 출력이 됨.
일반적으로 이 값 또는 top_p를 변경하는 것이 좋지만 둘 다 변경하는 것은 권장하지 않음.}

\item{top_p}{numeric. 온도를 이용한 샘플링의 대안으로, 핵 샘플링이라고 하며,
모델이 상위_p 확률 질량을 가진 토큰의 결과를 고려.
따라서 0.1은 상위 10\% 확률 질량을 구성하는 토큰만 고려한다는 의미.
일반적으로 이 값이나 temperature를 변경하는 것을 권장하지만 둘 다 변경하는 것은 권장하지 않음.}

\item{type}{character. 반환하는 결과 타입. "text", "console", "viewer"에서 선택하며,
기본값인 "text"는 텍스트를, "console"는 R 콘솔에 프린트 아웃되며,
"viewer"는 HTML 포맷으로 브라우저에 출력됨.}

\item{openai_api_key}{character. openai의 API key.}
}
\description{
chatGPT를 이용해서 Chat completion을 수행함.
}
\examples{
\dontrun{
# 텍스트로 반환
messages <- "mtcars 데이터를 ggplot2 패키지로 EDA하는 R 스크립트를 짜줘."
chat_completion(messages)

# R 콘솔에 프린트 아웃
chat_completion(messages, type = "console")

# HTML 포맷으로 브라우저에 출력
chat_completion(messages, type = "viewer")
}
}
