% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{add}
\alias{add}
\alias{create_messages}
\title{Create chat messages for chatGPT}
\usage{
add(messages, ...)

create_messages(user = NULL, system = NULL)
}
\arguments{
\item{user}{character. user role을 갖는 메시지.}

\item{system}{character. system role을 갖는 메시지.}
}
\value{
messages object.
}
\description{
chatGPT의 chat completion을 수행하기 위한 messages를 생성함
}
\examples{
\dontrun{
# 메시지 생성
msg <- create_messages("Who won the world series in 2020?", "You are a helpful assistant.")
msg

# 메시지 생성
msg <- create_messages("Who won the world series in 2020?")
msg
}
}
