% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{draw_img_edit}
\alias{draw_img_edit}
\title{Create image edit with chatGPT}
\usage{
draw_img_edit(
  image,
  mask,
  prompt,
  ko2en = TRUE,
  n = 1L,
  size = c("1024x1024", "256x256", "512x512"),
  type = c("url", "image", "file"),
  format = c("png", "jpeg", "gif"),
  path = "./",
  fname = "aiedit",
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{image}{character. 편집할 이미지 파일의 이름.}

\item{mask}{character. 투명한(alpha 값이 0인 경우) 영역이 이미지를 편집하 위치를 나타내는 추가 이미지.
4MB 미만의 유효한 PNG 파일이어야 하며 이미지와 크기가 같아야 함.}

\item{prompt}{character. 편집을 원하는 이미지에 대한 설명으로 최대 길이는 1000자.}

\item{ko2en}{logical. 프롬프트가 한국어일 때, 영어로 번역하여 질의하는 여부 설정.
TRUE이면 한글 프롬프트를 영어로 번역하여 프롬프트를 질의.}

\item{n}{integer. 생성할 이미지의 개수. 기본값은 1이며, 1과 10 사이의 정수를 사용함.}

\item{size}{character. 이미지의 크기로 "1024x1024", "256x256", "512x512"에서
하나를 선택함. 기본값은 "1024x1024".}

\item{type}{character. 반환하는 이미지 타입. "url", "image", "file"에서 선택하며,
기본값인 "url"은 이미지를 포함하는 URL을, "image"는 R환경에서 플롯으로 반환하고,
"image"는 이미지 파일을 생성합니다.}

\item{format}{character. 이미지 파일의 포맷으로 type의 값이 "file"일 경우만 적용됨.
"png", "jpeg", "gif"에서 선택하며, 기본값은 "png"임.}

\item{path}{character. 파일을 생성할 디렉토리 경로로 type의 값이 "file"일 경우만 적용됨.}

\item{fname}{character. 경로와 확장자를 제외한 이미지 파일의 이름으로,
type의 값이 "file"일 경우만 적용됨.}

\item{openai_api_key}{character. openai의 API key.}
}
\description{
chatGPT를 이용해서 이미지를 편집함.
}
\details{
편집할 이미지 파일은 정사각형의 png 파일만 지원하며, 파일의 용량은 4MB미만이어야 함.
}
\examples{
\dontrun{
# 편집할 이미지
image <- system.file("images", "cloud.png", package = "bitGPT")

# Mask 이미지
mask <- system.file("images", "cloud_mask.png", package = "bitGPT")

# 생성된 파일의 경로로서의 URL 반환
draw_img_edit(image, mask, prompt = "UFOs flying in the sky", ko2en = FALSE)

# 이미지를 반환
draw_img_edit(image, mask, prompt = "하늘을 날아다니는 UFO", type = "image")

# 파일로 출력
draw_img_edit(image, mask, prompt = "하늘을 날아다니는 UFO", type = "file")
}
}
\seealso{
\code{\link{draw_img}}, \code{\link{draw_img_variation}}
}
