% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{get_last_tokens}
\alias{get_last_tokens}
\title{Get usage token for last call}
\usage{
get_last_tokens(is_value = TRUE)
}
\arguments{
\item{is_value}{logical. TRUE이면 사용 토큰량들을 리스트 객체로 반환하고,
FALSE이면 그 정보를 콘솔에 출력합니다.}
}
\description{
마지막에 호출된 언어모델에 대한 사용 토큰 정보를 가져옵니다.
}
\details{
언어모델을 인터페이스한 "chat_completion()", "create_completion()", "create_embeddings()"만 지원합니다.
API key의 보안을 위해서 regist_naver_key()대신 set_naver_key()를 사용하세요.
}
\examples{
\dontrun{
poem <- c(
  "님은 갔습니다. 아아, 사랑하는 나의 님은 갔습니다. 푸른 산빛을 깨치고 단풍나무 숲을 향하야 난 적은 길을 걸어서 참어 떨치고 갔습니다.",
  "넓은 들 동쪽 끝으로 옛 이야기 지줄대는 실개천이 휘돌아 나가고 얼룩백이 황소가 해설피 금빛 게으른 울음을 우는 곳 그곳이 차마 꿈엔들 잊힐리야.",
  "한잔의 술을 마시고 우리는 버지니아울프의 생애와 목마를 타고 떠난 숙녀의 옷자락을 이야기한다.",
  "아아, 님은 갔지마는 나는 님을 보내지 아니하였습니다. 제 곡조를 못 이기는 사랑의 노래는 님의 침묵을 휩싸고 돕니다.",
  "전설바다에 춤추는 밤물결 같은 검은 귀밑머리 날리는 어린 누이와 아무렇지도 않고 예쁠 것도 없는 사철 발벗은 아내가 따가운 햇살을 등에 지고 이삭 줍던 곳, 그곳이 차마 꿈엔들 잊힐리야.",
  "목마는 주인을 버리고 거저 방울 소리만 울리며 가을 속으로 떠났다 술병에서 별이 떨어진다 상심한 별은 내 가슴에 가벼웁게 부서진다"
)

poem_embeddings <- create_embeddings(model = "text-embedding-ada-002", input = poem)

get_last_tokens()
}
}
