% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chating.R
\name{keep_completion}
\alias{keep_completion}
\title{Chat with chatGPT}
\usage{
keep_completion(prompt, last = TRUE, initialize = FALSE)
}
\arguments{
\item{prompt}{character. 채팅을 위한 user role의 프롬프트.}

\item{last}{logical. 마지막 user role의 프롬프트와 assistant role의 답변을 출력할지의 여부.
기본값은 TRUE로 마지막 대화내용만 출력함. FALSE일 경우에는 모든 대화 이력을 출력함.}

\item{initialize}{logical. 지존의 대화 정보를 초기화한 후, 채팅을 수할지의 여부.
기본값은 FALSE로 대화를 계속 이어감.}
}
\description{
chatGPT의 chat completion을 이용해서 채팅을 수행합니다.
}
\examples{
\dontrun{
# 대화 생성
keep_completion(prompt = "지금 대한민국에서 가장 중요한 이슈가 뭐야?")
keep_completion(prompt = "그렇다면 현재 진행 중인 상태는 어떤가?")

# 새로운 대화를 시작함
keep_completion(prompt = "지금 대한민국에서 가장 중요한 이슈가 뭐야?", initialize = TRUE)

# 처음 대화부터 출력함
keep_completion(prompt = "그렇다면 현재 진행 중인 상태는 어떤가?", last = FALSE)
}
}
\references{
{https://github.com/bit2r/bitGPT/issues/25 이 기능의 구현과 코드는 HelloDataScience님의 제안을 반영하였습니다. 감사드립니다.}
}
