% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{last.messages}
\alias{last.messages}
\title{Extract last messages from messages object}
\usage{
\method{last}{messages}(messages, ...)
}
\description{
messages 객체에서 마지막 대화를 추출합니다.
}
\details{
채팅을 기능을 구현하기 위해서 마지막 대화만 핸들링하는 경우에 유용합니다.
}
\examples{
\dontrun{
msg <- create_messages(user = "R을 이용한 통계학의 이해 커리큘럼을 부탁해",
                       system = "assistant는 R을 이용해서 통계학을 가르치는 강사입니다.")
show(msg)

# 메시지 객체로 반환
answer <- chat_completion(msg, type = "messages")
show(answer)

# 반환받은 메시지 객체에 질의를 위한 user role의 메시지 추가
msg <- add(answer, user = "커리큘럼에 tidyverse 패키지를 사용하는 방법을 추가해줘.")

# 이전 메시지를 포함하여 추가 질의
answer2 <- chat_completion(msg)
show(answer2)

last(answer2)
}
}
