% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{record_audio}
\alias{record_audio}
\title{Record audio}
\usage{
record_audio(
  time,
  sample_rate = 44100,
  channels = c("mono", "stereo"),
  fname = "voice.wav",
  overwrite = FALSE
)
}
\arguments{
\item{time}{numeric. 오디오를 레코딩할 시간. 단위는 초.}

\item{sample_rate}{integer. 샘플링 비율. 기본값은 44100으로 설정.}

\item{channels}{character. 레코딩 채널 설정. 모노와 스테레오 중에서 선택하며,
기본값은 모노인 "mono".}

\item{fname}{character. 레코딩된 오디오 파일의 이름. 기본값은 "voice.wav".}

\item{overwrite}{logical. 이미 오디오 파일이 존재할 때, 파일을 overwrite할지의 여부.}

\item{openai_api_key}{character. openai의 API key.}
}
\description{
record record audio using the current audio device
}
\details{
현재 오디오 드라이버의 오디오 인스턴스를 생성하여 오디오 녹음을 지정한 시간만큼 수행.
}
\examples{
\dontrun{
# 10초동안 오디오를 녹음함
record_audio(10)

# 10초동안 오디오를 스테레오로 녹음함
record_audio(10, channels = "stereo")
}
}
