% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{regist_deepl_key}
\alias{regist_deepl_key}
\title{Regist DeepL API key to package file}
\usage{
regist_deepl_key(api_key = NULL)
}
\arguments{
\item{api_key}{character. 등록할 DeepL API key.}
}
\description{
DeepL과 인터페이스하기 위한 DeepL API key를 등록합니다.
}
\details{
만약에 개인 컴퓨터가 아닌 여러 사용자가 사용하는 환경에 bitGPT 패키지를 설치한 경우라면,
API key의 보안을 위해서 regist_deepl_key()대신 set_deepl_key()를 사용하세요.
}
\examples{
\dontrun{
# 실제 사용자가 할당받은 DeepL API key를 사용합니다.
# regist_deepl_key("XXXXXXXXXXX")
}
}
