% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{retrieve_model}
\alias{retrieve_model}
\title{Retrieve models}
\usage{
retrieve_model(model)
}
\arguments{
\item{model}{character. 조회할 모델의 이름.}
}
\value{
a data.frame.
Names of data.frame is as follows.
\itemize{
\item id : 모델의 고유 식별자. 이 식별자는 모델을 호출하거나 다른 API 요청에서 참조할 때 사용됨.
\item object : 객체의 종류를 나타내는 문자열. 모든 "List models" API 결과의 object 값은 "model".
\item created : 모델이 생성된 날짜와 시간.
\item owned_by : 모델의 오너.
\item root : 모델의 루트.
\item parent : 모델의 부모 모델.
\item id : 권한의 고유 식별자.
\item object : 객체의 종류를 나타내는 문자열. 모든 object 값은 "model_permission".
\item created : 권한이 생성된 날짜와 시간.
\item allow_create_engine : 엔진 생성 가부.
\item allow_sampling : 샘플링 가부.
\item allow_logprobs : logprobs 가부.
\item allow_search_indices : search indices 가부.
\item allow_view : view 가부.
\item allow_fine_tuning : fine tuning 가부.
\item organization : organization.
\item group : 그룹.
\item is_blocking : 블록 여부.
}
}
\description{
모델 인스턴스를 검색하여 소유자 및 권한 부여와 같은 모델에 대한 기본 정보를 제공.
}
\examples{
\dontrun{
retrieve_model("text-davinci-002")
}
}
