% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{set_naver_key}
\alias{set_naver_key}
\title{Set Naver API key to system environment}
\usage{
set_naver_key(client_id = NULL, client_secret = NULL)
}
\arguments{
\item{client_id}{character. 등록할 API key의 client ID.}

\item{client_secret}{character. 등록할 API key의 client Secret.}
}
\description{
Naver 파파고 번역기와 인터페이스하기 위한 openai API key를 설정합니다.
}
\details{
만약에 여러 사용자가 사용하는 환경이 아닌 개인 컴퓨터에 bitGPT 패키지를 설치한 경우라면,
set_naver_key() 대신에 매번 API key를 등록할 필요없는 regist_naver_key()를 사용하세요.
}
\examples{
\dontrun{
# 실제 사용자가 할당받은 Naver API key를 사용합니다.
# set_naver_key()
}
}
