% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{set_openai_key}
\alias{set_openai_key}
\title{Set openai API key to system environment}
\usage{
set_openai_key(api_key = NULL)
}
\arguments{
\item{api_key}{character. 등록할 openai API key.}
}
\description{
openai와 인터페이스하기 위한 openai API key를 설정합니다.
}
\details{
만약에 여러 사용자가 사용하는 환경이 아닌 개인 컴퓨터에 bitGPT 패키지를 설치한 경우라면,
set_openai_key() 대신에 매번 API key를 등록할 필요없는 regist_openai_key()를 사용하세요.
}
\examples{
\dontrun{
# 실제 사용자가 할당받은 openai API key를 사용합니다.
# set_openai_key("XXXXXXXXXXX")
}
}
