% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{show.messages}
\alias{show.messages}
\title{Show chat messages object}
\usage{
\method{show}{messages}(messages, type = c("console", "viewer"), is_browse = TRUE, ...)
}
\arguments{
\item{messages}{messages. chatGPT와 chat completion messages 객체.}

\item{type}{character. 반환하는 결과 타입. console", "viewer"에서 선택하며,
기본값인 "console"는 R 콘솔에 프린트 아웃되며, "viewer"는 HTML 포맷으로 브라우저에 출력됨.
만약 결과에 R 코드가 chunk로 포함되어 있다면, 코드가 실행된 결과도 HTML 문서에 포함됨.}

\item{is_browse}{logical. type 인수가 TRUE일 경우에, is_browse가 TRUE이면
브라우저에 결과가 브라우징되고, FALSE이면 브라우징되지 않음.}
}
\description{
chatGPT의 chat completion를 위한 messages 객체를 조회함
}
\examples{
\dontrun{
msg <- create_messages(user = "R을 이용한 통계학의 이해 커리큘럼을 부탁해",
                       system = "assistant는 R을 이용해서 통계학을 가르치는 강사입니다.")
show(msg)

# 메시지 객체로 반환
answer <- chat_completion(msg, type = "messages")
show(answer)

# 반환받은 메시지 객체에 질의를 위한 user role의 메시지 추가
msg <- add(answer, user = "커리큘럼에 tidyverse 패키지를 사용하는 방법을 추가해줘.")

# 이전 메시지를 포함하여 추가 질의
answer2 <- chat_completion(msg)
show(answer2)
}
}
