% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{transcript_audio}
\alias{transcript_audio}
\title{Speech to Text with chatGPT}
\usage{
transcript_audio(
  file,
  language = "ko",
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{file}{character. 음성 오디오 파일 이름.}

\item{language}{character. 음성 오디오 파일의 언어. ISO-639-1 포맷으로 지정해야 하며,
기본값은 한국어인 "ko".}

\item{openai_api_key}{character. openai의 API key.}
}
\description{
chatGPT를 이용해서 음성 오디오 파일로 STT(Speech to Text)를 수행함.
}
\details{
오디오 파일의 포맷은 mp3, mp4, mpeg, mpga, m4a, wav, webm중 하나만 허용함.
}
\examples{
\dontrun{
# 음성 오디오 파일
speech <- system.file("audios", "korea_r_user.m4a", package = "bitGPT")

# 음성 오디오를 텍스트로 전환
transcript_audio(speech)
}
}
