% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{DeepL 번역기}
\usage{
translate(
  text = NULL,
  target = "EN",
  source = "KO",
  deepl_api_key = Sys.getenv("DEEPL_API_KEY")
)
}
\arguments{
\item{text}{character. 번역할 텍스트.}

\item{target}{character. 번역될 언어의 언어 코드. 기본값은 "EN"로 영어로 번역함.}

\item{source}{character. 번역할 텍스트 언어의 언어 코드. 기본값은 "KO"로 한국어를 번역함.}

\item{deepl_api_key}{character. DeepL API key.}
}
\description{
DeepL 번역 API를 이용해서 텍스트의 번역을 수행합니다.
}
\details{
지원되는 언어와 언어 코드는 https://developers.deepl.com/docs/api-reference/translate/openapi-spec-for-text-translation를 참고하세요.
}
\examples{
\dontrun{
text <- "빈센트 반 고흐 스타일로 일출과 갈매기를 그려줘"
translate(text)
}
}
