% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{update.messages}
\alias{update.messages}
\title{Update chat messages for chatGPT}
\usage{
\method{update}{messages}(
  messages,
  assistant = NULL,
  user = NULL,
  assistant_tokens = NULL,
  user_tokens = NULL,
  ...
)
}
\arguments{
\item{messages}{messages. chatGPT와 chat completion을 수행하기 위한 메시지 객체.}

\item{assistant}{character. assistant role을 갖는 메시지.}

\item{user}{character. user role을 갖는 메시지.}

\item{assistant_tokens}{integer. assistant 메시지로 산정된 completion tokens.}

\item{user_tokens}{integer. user 메시지로 산정된 prompt tokens.}
}
\value{
messages object.
}
\description{
chatGPT의 chat completion messages에 메시지와 토큰을 수정
}
\details{
API를 통한 호출과정에서 모델의 연산 결과로서의 토큰 사용량 정보를 업데이트하기 위한 용도입니다.
}
\examples{
\dontrun{
# 메시지 생성
msg <- create_messages("Who won the world series in 2020?", "You are a helpful assistant.")
msg

# 메시지 추가
library(dplyr)

msg <- msg \%>\%
  add(assistant = "The Los Angeles Dodgers won the World Series in 2020.",
      user = "Where was it played?")
msg

update(msg, user_tokens = 125)

}
}
