% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{draw_img}
\alias{draw_img}
\title{Create image with chatGPT}
\usage{
draw_img(
  prompt,
  ko2en = TRUE,
  n = 1L,
  size = c("1024x1024", "256x256", "512x512"),
  type = c("url", "image", "file"),
  format = c("png", "jpeg", "gif"),
  path = paste("aidrawing", format, sep = "."),
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{prompt}{character. 이미지 생성 명령을 수행할 프롬프트.}

\item{ko2en}{logical. 프롬프트가 한국어일 때, 영어로 번역하여 질의하는 것을 설정합니다.
TRUE이면 한글 프롬프트를 영어로 번역하여 프롬프트를 질의합니다.}

\item{n}{integer. 생성할 이미지의 개수. 기본값은 1이며, 0이상의 정수를 사용함.}

\item{size}{character. 이미지의 크기로 "1024x1024", "256x256", "512x512"에서
하나를 선택함. 기본값은 "1024x1024".}

\item{type}{character. 반환하는 이미지 타입. "url", "image", "file"에서 선택하며,
기본값인 "url"은 이미지를 포함하는 URL을, "image"는 R환경에서 플롯으로 반환하고,
"image"는 이미지 파일을 생성합니다.}

\item{format}{character. 이미지 파일의 포맷으로 type의 값이 "file"일 경우만 적용됨.
"png", "jpeg", "gif"에서 선택하며, 기본값은 "png"임.}

\item{path}{character. 경로를 포함한 이미지 파일의 이름으로 type의 값이 "file"일 경우만 적용됨.}

\item{openai_api_key}{character. openai의 API key.}
}
\description{
chatGPT를 이용해서 프로프트에 부합하는 이미지를 생성함.
}
\details{
openai의 프롬프트를 한글로 만들 경우에는 결과의 성능이 매우 낮기 때문에,
한글 프롬프트의 경우에는 ko2en를 TRUE로 지정하기 바랍니다.
이 경우에는 네이버의  파파고 번역 결과로 질의합니다.
}
\examples{
\dontrun{
# 영어 프롬프트
prompt_en <- "Draw a sunrise and a seagull in Vincent van Gogh style."
draw_img(prompt_en, ko2en = FALSE)

# 한글 프롬프트
prompt_ko <- "빈센트 반 고흐 스타일로 일출과 갈매기를 그려줘"
draw_img(prompt_ko, type = "image")

# 파일로 출력
draw_img(prompt_ko, type = "file")
}
}
