% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Naver 파파고 번역기}
\usage{
translate(
  text = NULL,
  source = "ko",
  target = "en",
  client_id = Sys.getenv("NAVER_CLIENT_ID"),
  client_secret = Sys.getenv("NAVER_CLIENT_SECRET")
)
}
\arguments{
\item{text}{character. 번역할 텍스트.}

\item{source}{character. 번역할 텍스트 언어의 언어 코드. 기본값은 "ko"로 한국어를 번역함.}

\item{target}{character. 번역될 언어의 언어 코드. 기본값은 "en"로 영어로 번역함.}

\item{client_id}{character. 등록할 API key의 client ID.}

\item{client_secret}{character. 등록할 API key의 client Secret.}
}
\description{
Naver 파파고 번역 API를 이용해서 텍스트의 번역을 수행합니다.
}
\details{
지원되는 언어와 언어 코드는 https://developers.naver.com/docs/papago/papago-nmt-api-reference.md를 참고하세요.
}
\examples{
\dontrun{
text <- "빈센트 반 고흐 스타일로 일출과 갈매기를 그려줘"
translate(text)
}
}
